feat: refactoring architecture of project to modularize files
