+++ .gitignore
@@ -1,4 +1,3 @@
 .venv
 .vscode
 __pycache__
-*.log
